refactor: move coverage tests into data
