+++ packages/dressed/src/resources/make/data.test.ts
@@ -1,7 +1,42 @@
 import { describe, expect, test } from "bun:test";
-import { routeKeyToMethodName } from "./data.ts";
+import { Routes } from "discord-api-types/v10";
+import routeDefinitions, { routeKeyToMethodName } from "./data.ts";
 
-const testRoutes: Record<string, Parameters<typeof routeKeyToMethodName>> = {
+const excused = new Set([
+  // Deprecated:
+  "channelPin",
+  "guildCurrentMemberNickname",
+  "nitroStickerPacks",
+  // Only sent from Slack/Github:
+  "webhookPlatform",
+  // Used, but not recognized:
+  "channelJoinedArchivedThreads",
+  "channelPins",
+  "channelThreads",
+  "guildMFA",
+  "guilds",
+  "interactionCallback",
+  "soundboardDefaultSounds",
+  // TODO:
+  "guildMemberVerification", // Documentation removed temporarily and API changing
+  "oauth2CurrentApplication",
+  "oauth2CurrentAuthorization",
+  "oauth2Authorization",
+  "oauth2TokenExchange",
+  "oauth2TokenRevocation",
+]);
+
+test("should have no missing and unexcused routes", () => {
+  const unused = new Set(Object.keys(Routes));
+  for (const [key, { overrides: { apiRoute } = {} }] of Object.entries(routeDefinitions)) {
+    const method = (key.match(/[A-Z][a-z]+/) ?? [])[0] ?? "";
+    const routeKey = apiRoute ?? key.slice(method.length).replace("API", "");
+    unused.delete(`${routeKey[0].toLowerCase()}${routeKey.slice(1)}`);
+  }
+  expect(Array.from(unused).filter((r) => !excused.has(r))).toBeEmpty();
+});
+
+const routeCases: Record<string, Parameters<typeof routeKeyToMethodName>> = {
   listRoutes: ["Get", "GetAPIRoutes"],
   getUser: ["Get", "GetAPIUser"],
   listRoles: ["Get", "GetAPIGuildRoles"],
@@ -14,7 +49,7 @@ const testRoutes: Record<string, Parameters<typeof routeKeyToMethodName>> = {
 };
 
 describe("routeKeyToMethodName", () => {
-  for (const [expected, params] of Object.entries(testRoutes)) {
+  for (const [expected, params] of Object.entries(routeCases)) {
     test(expected, () => {
       expect(routeKeyToMethodName(...params)).toBe(expected);
     });
